+++ pytorch_lightning/loops/epoch/training_epoch_loop.py
@@ -480,7 +480,10 @@ class TrainingEpochLoop(loops.Loop[_OUTPUTS_TYPE]):
         if not self.trainer.enable_validation:
             return False
 
-        is_val_check_epoch = (self.trainer.current_epoch + 1) % self.trainer.check_val_every_n_epoch == 0
+        is_val_check_epoch = (
+            self.trainer.check_val_every_n_epoch is None
+            or (self.trainer.current_epoch + 1) % self.trainer.check_val_every_n_epoch == 0
+        )
         if not is_val_check_epoch:
             return False
 
@@ -492,7 +495,11 @@ class TrainingEpochLoop(loops.Loop[_OUTPUTS_TYPE]):
         if self.trainer.should_stop:
             return True
 
-        # TODO(@awaelchli): let training/eval loop handle logic around limit_*_batches and val_check_batch
+        if self.trainer.check_val_every_n_epoch is None:
+            return (self.trainer.global_step + 1) % self.trainer.val_check_batch == 0
+
+        # TODO(@awaelchli): let training/eval loop handle logic around limit_*_ba
+        #  tches and val_check_batch
         is_val_check_batch = is_last_batch
         if isinstance(self.trainer.limit_train_batches, int) and is_infinite_dataset:
             is_val_check_batch = (batch_idx + 1) % self.trainer.limit_train_batches == 0